[history,shell] Split ores.history into api/core, add generic diff renderer
Restructure ores.history to the api/core layout used elsewhere in the
codebase, and add the shell's generic unified-diff history renderer
that consumes it.

Story-ID: 037B474D-84ED-4888-9054-D58AB1FB10D2
Task-ID: E4DDC1B9-8609-4C11-9172-F2EE79242E66
