+++ doc/agile/versions/v0/sprint_23/consolidate_history_dialogs/story.org
@@ -164,7 +164,7 @@ that rollout produces.
 | [[id:15CFB0BE-C411-4B12-B733-974F6F5A44A5][Generic history.v1.get NATS subject and server-side dispatch registry]] | DONE | 2026-07-12 | 2026-07-13 | Add the generic history.v1.get NATS subject and request/response types (entity_type, entity_id -> version/fields/changes), a server-side dispatch registry keyed by entity_type_of() reapplying the entity-composed-registrars meta-pattern, and register each entity's history field mapper as a provider. Remove the per-entity get_<entity>_history subjects/handlers as each entity migrates. |
 | [[id:7D8659AB-220E-4077-BD17-3C4DCA1937D8][Single generic HistoryDialog widget in ores.qt]] | BACKLOG | | | Build the single, non-templated HistoryDialog widget (version list, async load, toolbar, diff_span colour-highlighted changes tab per the GitHub-style capture) taking (entity_type, entity_id); add the codegen'd per-entity Open/Revert action registry; wire callers to construct HistoryDialog instead of a typed dialog subclass; wire the composite-entity 'Details' affordance for child-driven versions to open another HistoryDialog instance. |
 | [[id:CB042B99-9B2F-46E9-9CF5-73B47185BCD1][Retire HistoryDialogBase, per-entity dialogs, and their templates]] | BACKLOG | | | Once every entity has migrated onto the generic HistoryDialog, delete HistoryDialogBase, every remaining per-entity *HistoryDialog class, and cpp_qt_history_dialog.{hpp,cpp}.mustache/qt_history_dialog_ui.mustache; confirm via zero-diff regeneration guardrails that no facet still emits a per-entity dialog class. |
-| [[id:E4DDC1B9-8609-4C11-9172-F2EE79242E66][Generic unified-diff history renderer in ores.shell]] | BACKLOG | | | Replace the disabled currencies history-diff command with one generic, entity-parameterised unified-diff renderer consuming the generic history.v1.get response (context lines from fields, +/- pairs with span-based highlighting from changes); re-enable it for every entity; update the shell recipes. |
+| [[id:E4DDC1B9-8609-4C11-9172-F2EE79242E66][Generic unified-diff history renderer in ores.shell]] | STARTED | 2026-07-13 | | Replace the disabled currencies history-diff command with one generic, entity-parameterised unified-diff renderer consuming the generic history.v1.get response (context lines from fields, +/- pairs with span-based highlighting from changes); re-enable it for every entity; update the shell recipes. |
 | [[id:1EAE3E54-1943-46DA-AB9A-FFC9D26287E6][Refactor history dialog UI per Gemini review]] | BACKLOG | | | Redesign HistoryDialogBase per doc/analysis/gemini_review_history_dialog.org: master-detail split with a live side-by-side diff panel (semantic colour highlighting, inline word diffing), multi-version compare mode, and a compressed timeline with version pill badges, replacing the current flat Old Value/New Value table and dead right-hand space. |
 
 * Decisions
