+++ projects/CMakeLists.txt
@@ -63,7 +63,7 @@ endif()
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.platform)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.logging)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.diff)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.history)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.history/api)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.orgmode)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.database)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.testing)
@@ -81,6 +81,7 @@ add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.synthetic/api)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.synthetic/core)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.security)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.service)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.history/core)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.geo)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.iam/api)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.iam/client)
