+++ projects/ores.history/core/modeling/ores.history.core.puml
@@ -1,0 +1,53 @@
+' -*- mode: plantuml; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
+'
+' Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
+'
+' This program is free software; you can redistribute it and/or modify it under
+' the terms of the GNU General Public License as published by the Free Software
+' Foundation; either version 3 of the License, or (at your option) any later
+' version.
+'
+' This program is distributed in the hope that it will be useful, but WITHOUT
+' ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+' FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+'
+' You should have received a copy of the GNU General Public License along with
+' this program; if not, write to the Free Software Foundation, Inc., 51 Franklin
+' Street, Fifth Floor, Boston, MA 02110-1301, USA.
+'
+@startuml
+
+title ores.history.core Component
+
+set namespaceSeparator ::
+
+namespace ores #F2F2F2 {
+    namespace history #F2F2F2 {
+        namespace messaging #FEFECE {
+            class history_handler <<class>> {
+                +history(msg) : void
+            }
+            history_handler --> service::dispatch_registry : dispatches via
+            history_handler ..> ores::database::context : resolves via make_request_context()
+        }
+        namespace service #FEFECE {
+            class dispatch_registry <<class>> {
+                +register_history_provider(entity_type, provider) : void
+                +has_provider(entity_type) : bool
+                +provider_count() : size_t
+                +dispatch(request, ctx) : get_entity_history_response
+            }
+        }
+    }
+}
+
+note right of ores::history::messaging::history_handler
+  register_history_handlers(nats, registry, queue_group, ctx, verifier)
+  subscribes the single history.v1.get subject to a
+  history_handler backed by the caller-owned registry.
+  Each request is resolved into a scoped database::context
+  via make_request_context(), exactly like every other NATS
+  handler in the composing service — no bespoke auth channel.
+end note
+
+@enduml
