+++ projects/ores.history/core/src/CMakeLists.txt
@@ -1,6 +1,6 @@
 # -*- mode: cmake; cmake-tab-width: 4; indent-tabs-mode: nil -*-
 #
-# Copyright (C) 2025 Marco Craveiro <marco.craveiro@gmail.com>
+# Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
 #
 # This program is free software; you can redistribute it and/or modify it under
 # the terms of the GNU General Public License as published by the Free Software
@@ -9,13 +9,14 @@
 #
 # This program is distributed in the hope that it will be useful, but WITHOUT
 # ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.
 #
 # You should have received a copy of the GNU General Public License along with
-# this program; if not, write to the Free Software Foundation, Inc., 51 Franklin
-# Street, Fifth Floor, Boston, MA 02110-1301, USA.
+# this program; if not, write to the Free Software Foundation, Inc., 51
+# Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 #
-set(name "ores.refdata.core")
+set(name "ores.history.core")
 set(lib_binary_name ${name})
 set(lib_target_name ${name}.lib)
 
@@ -26,35 +27,22 @@ file(GLOB_RECURSE files RELATIVE
 
 set(lib_files ${files})
 add_library(${lib_target_name} ${lib_files})
-target_compile_definitions(${lib_target_name} PRIVATE ORES_REFDATA_CORE_LIBRARY)
-if(CMAKE_CXX_COMPILER_ID MATCHES "GNU|Clang")
-    target_compile_options(${lib_target_name} PRIVATE
-        -fvisibility=hidden -fvisibility-inlines-hidden)
-endif()
+target_compile_definitions(${lib_target_name} PRIVATE ORES_HISTORY_CORE_LIBRARY)
 set_target_properties(${lib_target_name} PROPERTIES
     OUTPUT_NAME ${lib_binary_name})
 set_target_properties(${lib_target_name}
     PROPERTIES VERSION ${PROJECT_VERSION} SOVERSION ${PROJECT_VERSION_MAJOR})
 
 target_include_directories(${lib_target_name} PUBLIC
-    ${CMAKE_SOURCE_DIR}/projects/ores.refdata/core/include)
+    ${CMAKE_SOURCE_DIR}/projects/ores.history/core/include)
 
 target_link_libraries(${lib_target_name}
     PUBLIC
-        ores.refdata.api.lib
+        ores.history.api.lib
         ores.database.lib
+        ores.logging.lib
         ores.nats.lib
-        ores.diff.lib
-    PRIVATE
-        ores.dq.api.lib
-        ores.history.core.lib
-        ores.service.lib
-        pugixml::pugixml
-        ores.variability.api.lib
-        ores.platform.lib
-        ores.utility.lib
         ores.security.lib
-        faker-cxx::faker-cxx
-        libfort::fort)
+        ores.service.lib)
 
 install(TARGETS ${lib_target_name} LIBRARY DESTINATION lib)
