--- projects/ores.history/include/ores.history/export.hpp
@@ -1,31 +1,0 @@
-/* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
- *
- * Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 3 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
- */
-#ifndef ORES_HISTORY_EXPORT_HPP
-#define ORES_HISTORY_EXPORT_HPP
-
-#include <boost/config.hpp>
-
-#ifdef ORES_HISTORY_LIBRARY
-#    define ORES_HISTORY_EXPORT BOOST_SYMBOL_EXPORT
-#else
-#    define ORES_HISTORY_EXPORT BOOST_SYMBOL_IMPORT
-#endif
-
-#endif
